requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
cv==1.0.0
Django==3.1
django-ckeditor==5.7.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.9.1
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
oscrypto==1.3.0
packaging==21.3
pdf2image==1.16.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyparsing==3.0.9
PyPDF3==1.0.6
pytesseract==0.3.9
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.11
webencodings==0.5.1
xhtml2pdf==0.2.8
